Bump to 1.68, fix Clippy warnings, change build
Ultimate goal: bumping toolchain rev to 1.68.

The continuous build was using a separate notion of MSRV (which happened
to equal the pinned toolchain but needed to be updated separately), so
the original bump didn't pass. I have rearranged that build after
talking with @steveklabnik -- the original intent was to build on both
pinned and latest-stable, and it now does that without needing a
separate MSRV variable.

I have, however, added an MRSV to Cargo.toml to give better feedback to
people who try and build Humility with a too-old toolchain.

Once I got past that hurdle, I hit the build that requires Humility to
be clippy clean with the pinned toolchain -- and evidently it wasn't.
The things it reported were absolutely lints in previous versions of
clippy -- I guess they got missed because of some bugs in clippy that
got fixed? Anyway, I have gone through and fixed them all, generally in
the most minimal way possible, but making a related nearby cleanup in a
few cases to simplify the code.

This seems to work now.

Note that the 1.68 toolchain is now warning us about some deps
containing deprecated code, which we should probably fix.

+++ cmd/spctrl/src/lib.rs
@@ -108,7 +108,7 @@ fn spctrl(context: &mut humility::ExecutionContext) -> Result<()> {
                     bail!("invalid byte {}", byte)
                 }
             }
-            ops.push(Op::Push32(addr as u32));
+            ops.push(Op::Push32(addr));
             ops.push(Op::Push32(arr.len() as u32));
 
             let sp_write = funcs.get("WriteToSp", 2)?;
@@ -120,7 +120,7 @@ fn spctrl(context: &mut humility::ExecutionContext) -> Result<()> {
             println!("{:x?}", results);
         }
         SpCtrlCmd::Read { addr, nbytes } => {
-            ops.push(Op::Push32(addr as u32));
+            ops.push(Op::Push32(addr));
             ops.push(Op::Push32(nbytes as u32));
             let sp_read = funcs.get("ReadFromSp", 2)?;
             ops.push(Op::Call(sp_read.id));
